Migrate to use tools-deps instead of lein
We (Yummly) is having issues with main line not having a regular release cycle and
don't want to continue to maintain a private maven repository or regularly release
to a yummly specific version of manifold on clojars. So instead, just migrate to tools-deps
so we can rely on the code by git sha.