Add core.async/go variant that works with defferreds
`let-flow` explicitly expects that everything inside of it should be resolved before
passed to other functions. Instead, let's introduce a new macro that still allows for
normal looking clojure code, but adds ability to park when specified. Instead of writing
our own code for this, let's take advantage of the fact that core.async's state-machine code
is quite reusable.